fix kernel codes

+++ arch/arm/mach-msm/qdsp6v2/apr.c
@@ -59,7 +59,7 @@ int apr_send_pkt(void *handle, uint32_t *buf)
 	struct apr_hdr *hdr;
 	uint16_t dest_id;
 	uint16_t client_id;
-	int w_len;
+	uint16_t w_len;
 	unsigned long flags;
 
 	if (!handle || !buf) {
